add isAddParticipant flag
