+++ app/src/main/java/com/nextcloud/talk/contacts/SearchComponent.kt
@@ -14,11 +14,13 @@ import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.Close
+import androidx.compose.material3.Button
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.res.stringResource
@@ -30,6 +32,7 @@ import com.nextcloud.talk.R
 
 @Composable
 fun DisplaySearch(text: String, onTextChange: (String) -> Unit, contactsViewModel: ContactsViewModel) {
+    val isAddParticipants = contactsViewModel.isAddParticipantsView.collectAsState()
     val keyboardController = LocalSoftwareKeyboardController.current
     TextField(
         modifier = Modifier
@@ -42,7 +45,6 @@ fun DisplaySearch(text: String, onTextChange: (String) -> Unit, contactsViewMode
                 text = stringResource(R.string.nc_search)
             )
         },
-
         textStyle = TextStyle(
             fontSize = 16.sp
         ),
@@ -91,4 +93,10 @@ fun DisplaySearch(text: String, onTextChange: (String) -> Unit, contactsViewMode
         ),
         maxLines = 1
     )
+    if (isAddParticipants.value) {
+        Button(onClick = {
+        }, modifier = Modifier.fillMaxWidth(0.2f)) {
+            Text(text = "Done")
+        }
+    }
 }
